surveyhubbackend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.10.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==47.0.0
Django==4.2.25
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
google-auth==2.49.2
google-genai==1.47.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.13.3
pydantic_core==2.46.3
PyJWT==2.12.1
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.3
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==15.0.1
gunicorn==21.2.0
